implemented end-to-end testing for main requirements
In this commit, I also change the way validation pipe is initiated globally with the use of custom validation pipe and insert it as a provider in App's module, this is for end-to-end testing to enable DI for the pipe (because previous useGlobalPipes is not working for e2e test). Other than that, i also updates the DTOs for creating and updating watch at release_date column. This update makes watch service to not have to convert it to js Date before throw it into prisma.